Add lunar eclipse sequences.
A bunch of different versions of the sequence. Some using auto-exposure and others without it. If all runs perfectly I could just use one of the full-event sequences (Lunar_Eclipse.scs or Lunar_Eclipse_AutoExposure.scs). If not, I've got the simpler sequences to run. Though I'll need to keep track of the timing myself for those.